Don't use gulp for running tests
npm is more than enough to instantiate intern with the correct args and
removes a bunch of complexity.
