+++ gulpfile.babel.js
@@ -5,7 +5,6 @@ import del from 'del';
 import source from 'vinyl-source-stream';
 import buffer from 'vinyl-buffer';
 import gulp from 'gulp';
-import childProcess from 'child_process';
 import oghliner from 'oghliner';
 import loadPlugins from 'gulp-load-plugins';
 const plugins = loadPlugins({
@@ -26,16 +25,6 @@ const browserSync = browserSyncCreator.create();
 
 import engine from './engine/index.js';
 
-import https from 'https';
-
-const testsVarDir = 'tests/support/var/';
-const seleniumVersion = '2.48.2';
-const seleniumFilename = 'selenium-server-standalone-' + seleniumVersion + '.jar';
-const seleniumPath = testsVarDir + seleniumFilename;
-const seleniumDownloadUrl = 'https://selenium-release.storage.googleapis.com/2.48/selenium-server-standalone-2.48.2.jar';
-const seleniumLogPath = testsVarDir + 'selenium.log';
-
-
 gulp.task('clean', () => {
   return del(['./dist']);
 });
@@ -46,103 +35,6 @@ gulp.task('lint', () => {
     .pipe(plugins.eslint())
     .pipe(plugins.eslint.format());
 });
-
-gulp.task('test:node', () => {
-  return new Promise((resolve, reject) => {
-    const child = childProcess.spawn('./node_modules/intern/bin/intern-client.js', ['config=tests/intern-node'], { stdio: 'inherit' });
-    child.once('exit', (exitCode) => {
-      if (exitCode !== 0) {
-        reject();
-      } else {
-        resolve();
-      }
-    });
-  });
-});
-
-function maybeMkdir(path) {
-  return new Promise((resolve, reject) => {
-    // We ignore creation errors since we don't care
-    // whether the directory was already there or if
-    // we created it, and we're about to check its
-    // existence.
-    fs.mkdir(path, () => {
-      fs.stat(path, (err, stats) => {
-        if (err) {
-          reject(err);
-          return;
-        }
-
-        if (!stats.isDirectory()) {
-          reject(new Error(path + ' already exists and is not a directory'));
-          return;
-        }
-
-        resolve();
-      });
-    });
-  });
-}
-
-function ensureSelenium() {
-  return new Promise((resolve, reject) => {
-    // Check for selenium server JAR
-    fs.access(seleniumPath, fs.R_OK, (accessErr) => {
-      if (!accessErr) {
-        console.log('Using existing Selenium server JAR');
-        resolve();
-        return;
-      }
-
-      console.log('No Selenium server JAR found');
-      // Download Selenium if not found
-      maybeMkdir(testsVarDir).then(() => {
-        const file = fs.createWriteStream(seleniumPath);
-        const request = https.get(seleniumDownloadUrl, (response) => {
-          console.log('Downloading Selenium server JAR');
-          response.pipe(file);
-          file.on('finish', resolve);
-        });
-
-        request.on('error', (reqErr) => { // Handle errors
-          fs.unlink(seleniumPath); // Delete the file async. (But we don't check the result)
-          reject(reqErr);
-        });
-      });
-    });
-  });
-}
-
-gulp.task('test:browser', ['build'], () => {
-  return ensureSelenium().then(() => {
-    return new Promise((resolve, reject) => {
-      fs.open(seleniumLogPath, 'w', (err, fd) => {
-        if (err) {
-          reject(err);
-          return;
-        }
-
-        const server = childProcess.spawn('java', ['-jar', seleniumPath], { stdio: [fd, fd, fd] });
-
-        // Wait 1s after starting the Selenium server to give it time
-        // to start accepting connections
-        setTimeout(() => {
-          const child = childProcess.spawn('./node_modules/intern/bin/intern-runner.js', ['config=tests/intern-browser'], { stdio: 'inherit' });
-          child.once('exit', (exitCode) => {
-            server.kill('SIGINT');
-            if (exitCode !== 0) {
-              reject();
-            } else {
-              resolve();
-            }
-          });
-        }, 1000);
-      });
-    });
-  });
-});
-
-gulp.task('test', ['lint', 'test:node', 'test:browser']);
 
 gulp.task('deploy', ['build'], () => {
   return oghliner.deploy({
