+++ package.json
@@ -40,7 +40,8 @@
     "yargs": "^3.27.0"
   },
   "scripts": {
-    "test": "./node_modules/gulp/bin/gulp.js test",
+    "lint": "gulp lint",
+    "test": "npm run lint && node tests/support/runtests.js",
     "start": "./node_modules/gulp/bin/gulp.js watch"
   },
   "repository": {
